chore(definition): apply umd
