+++ types/test-types.ts
@@ -1,4 +1,4 @@
-import feathers from './';
+import feathers from '../';
 
 const app = feathers();
 
@@ -9,6 +9,8 @@ const service = app.service<model>('user');
 (async () => {
   let m1 = await service.get(1);
   let m2 = await service.patch(m1.id, {user:''});
-
-  console.log(m2.pass);
+  
+  service.on('data', () =>{
+    console.log(m2.pass);
+  });
 })();
